WIP: Profile Delete and Settings Dialog
Starting to have more windows. Will I need UIManager?
Need to format credits properly. Would be nice to animate the scroll but it was pretty complicated to get right in Solar Rogue so maybe a waste of time.
Need to scale the slide bar for volume but not sure of how to do it properly.